requirements.txt
numpy
pandas
PyQt5
pyqtgraph
pyserial
opencv-python